requirements.txt
chromadb
openai
python-dotenv
flask
gradio